requirements.txt
# 核心依赖
pyyaml>=5.1
# 可选依赖
Flask>=2.0.0
Django>=3.2.0